requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
